refactor: removed unused component
